Needed to calculate using double - using int directly caused some problems with how the values were being operated on. Also limited some values to 2 decimal places, as they were shockingly long

+++ Assig1/Controllers/ExpiationsController.cs
@@ -74,7 +74,7 @@ namespace Assig1.Controllers
             var index = _context.Expiations
                 .Count(e => e.VehicleSpeed.HasValue && e.VehicleSpeed <= expiation.VehicleSpeed);
 
-            var speedPercentile = index / speedCount * 100;
+            var speedPercentile = (double)index / speedCount * 100;
 
             var vm = new Expiation_ExpiationDetail
             {
